edge_agent/requirements.txt
httpx==0.27.0
pyserial==3.5
pydantic-settings==2.3.0
